requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.26.0
pillow>=10.0.0
requests>=2.31.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.2
sqlalchemy>=2.0.20

torch==2.2.0+cpu
torchvision==0.17.0+cpu
